이미지, 비디오 업로드 API Riot -> Post로 이동 (#74)
