+++ src/main/java/com/gamemoonchul/infrastructure/web/RiotApiController.java
@@ -13,23 +13,9 @@ import org.springframework.web.multipart.MultipartFile;
 @RequestMapping("/api/riot")
 public class RiotApiController {
     private final RiotApiService riotApiService;
-    private final S3Service s3Service;
 
     @GetMapping("/search-match/{gameId}")
     public MatchGameResponse searchMatch(@PathVariable String gameId) {
         return riotApiService.searchMatch("KR_" + gameId);
     }
-
-    @PostMapping("/video-upload")
-    public String uploadVideo(@RequestParam("file") MultipartFile file) {
-        String fileUrl = s3Service.uploadVideo(file);
-        return fileUrl;
-    }
-
-    @PostMapping("/thumbnail-upload")
-    public String uploadThumbnail(@RequestParam("file") MultipartFile file) {
-        String fileUrl = s3Service.uploadImage(file);
-        return fileUrl;
-    }
-
 }
